// marketplace/src/types/custom-interface.ts
export interface IPropertyReviewFormData {
  name: string;
  email: string;
  number: string;
  message: string;
  termsAccepted: boolean;
}